+++ forma/init.lua
@@ -2,9 +2,10 @@
 -- The more rigourous procedure of
 --  local cell = require('forma.cell')
 -- is recommended.
-require('forma.cell')
-require('forma.pattern')
-require('forma.primitives')
-require('forma.subpattern')
-require('forma.automata')
-require('forma.neighbourhood')
+cell          = require('forma.cell')
+pattern       = require('forma.pattern')
+primitives    = require('forma.primitives')
+subpattern    = require('forma.subpattern')
+automata      = require('forma.automata')
+neighbourhood = require('forma.neighbourhood')
+