+++ forma/pattern.lua
@@ -1,18 +1,22 @@
 --- A class containing a set or *pattern* of cells.
 --
--- The **pattern** class is the central class of this module, representing a
--- set of `cell`s. This set can be initialised as empty, or according to
--- geometric `primitives`. Once initialised, a pattern can only be modified by
+-- The **pattern** class is the central class of `forma`, representing a set of
+-- points or *cells*. This set can be initialised as empty, or according to a
+-- 'prototype' consisting of a NxM table matrix of 1's or 0's. Several helper
+-- methods for the initialisation of a `pattern` are provided in the
+-- `primitives` module.  Once initialised, a pattern can only be modified by
 -- the `insert` method, used to add active cells. All other pattern
 -- manipulations return a new, modified pattern rather than modifying patterns
 -- in-place.
 --
--- Several pattern manipulators are provided here. For example as a `shift` of
--- an entire pattern, manipulators that `enlarge` a pattern by a scale factor
--- or performing reflections in the x or y axes. Particuarly useful are
--- manipulators which generate new patterns as the `edge` (outer hull) or
--- `surface` (inner-hull) of other patterns. These manipulators can be used
--- with different definitions of a cell's `neighbourhood`.
+-- Several pattern manipulators are provided here. For example a `shift`
+-- manipulator which shifts the coordinates of an entire pattern, manipulators
+-- that `enlarge` a pattern by a scale factor and modifiers than can `rotate`
+-- or reflect patterns in the x (`hreflect`) or y (`vreflect`) axes.
+-- Particuarly useful are manipulators which generate new patterns such as the
+-- `edge` (outer hull) or `surface` (inner-hull) of other patterns. These
+-- manipulators can be used with custom definitions of a cell's
+-- `neighbourhood`.
 --
 -- Pattern coordinates should be reliable in [-65536, 65536]. This is
 -- adjustable through the `MAX_COORDINATE` constant.
@@ -29,6 +33,11 @@
 --
 -- -- 'Method' style with chaining used for :insert
 -- local p2 = pattern.new():insert(1,1) -- Idential as to p1
+--
+-- -- 'Prototype' style
+-- local p3 = pattern.new({{1,1,1},
+--                         {1,0,1},
+--                         {1,1,1}})
 --
 -- -- Fetch a random cell and the medoid (centre-of-mass) cell from a pattern
 -- local random_cell = p1:rcell()
@@ -70,21 +79,37 @@ local function key_to_coordinates(key)
     return xp- MAX_COORDINATE, yp - MAX_COORDINATE
 end
 
+-- Copies an input table in a randomly shuffled order
+local function shuffled_copy(intable, rng)
+    -- Copy and Fisher-Yates shuffle
+    local extable = {}
+    for i = 1, #intable, 1 do
+        local j = rng(1, i)
+        if j ~= i then
+            extable[i] = extable[j]
+        end
+        extable[j] = intable[i]
+    end
+    return extable
+end
 
 --- Basic methods.
 -- Methods for the creation, copying and adding of cells to a pattern.
 --@section basic
 
 --- Pattern constructor.
--- Points are stored in the pattern in a standard integer keyed table, and
--- also as elements in a spatial hash map.
--- @param prototype (optional) an N*N 2D table of ones and zeros.
--- Returns a new pattern. If no prototype is used, then an empty pattern
--- is returned. If set with the prototype table {{1,0},{0,1}} will initialise
--- the pattern:
---  10
---  01
--- @return new pattern
+--  This method returns a new pattern, according to a prototype. If no
+--  prototype is used, then an empty pattern is returned. For example, if
+--  called with the prototype `{{1,0},{0,1}}` this method will return the
+--  pattern:
+--  `
+--    10
+--    01
+--  `
+-- Active cells are stored in the pattern in a standard integer keyed table, and
+-- also as elements in a spatial hash map for fast look-up of active cells.
+-- @param prototype (optional) an N*M 2D table of ones and zeros
+-- @return a new pattern according to the prototype
 function pattern.new(prototype)
     local np = {}
 
@@ -103,17 +128,21 @@ function pattern.new(prototype)
     if prototype ~= nil then
         assert(type(prototype) == 'table',
         'pattern.new requires either no arguments or a N*N matrix as a prototype')
-        local N = #prototype
-        for i=1,N,1 do
+        local N, M = #prototype, nil
+        for i=1, N, 1 do
             local row = prototype[i]
             assert(type(row) == 'table',
             'pattern.new requires either no arguments or a N*N matrix as a prototype')
-            assert(#row == N,
-            'pattern.new requires a N*N matrix when using a prototype, you requested '..N..'*'.. #row)
-            for j=1,N,1 do
+            if i == 1 then
+                M = #row
+            else
+                assert(#row == M,
+                'pattern.new requires a N*M matrix as prototype when called with an argument')
+            end
+            for j=1, M, 1 do
                 local icell = row[j]
                 if icell == 1 then
-                    np:insert(i-1,j-1) -- Patterns start from zero
+                    np:insert(j-1,i-1) -- Patterns start from zero
                 else
                     assert(icell == 0, 'pattern.new: invalid prototype entry (must be 1 or 0): '.. icell)
                 end
@@ -126,7 +155,7 @@ end
 
 --- Copy an existing pattern.
 -- @param ip input pattern for cloning
--- @return A copy of the pattern ip
+-- @return a copy of the pattern ip
 function pattern.clone(ip)
     assert(getmetatable(ip) == pattern, "pattern cloning requires a pattern as the first argument")
     local newpat = pattern.new()
@@ -209,16 +238,19 @@ end
 --- Iterators.
 -- @section Iterators
 
---- Return an iterator over active cells in the pattern.
+--- Iterator over active cells in the pattern.
 -- @param ip source pattern for active cell iterator
--- @return an iterator providing a `cell` table for every active cell in the pattern
+-- @return an iterator returning a `cell` for every active cell in the pattern
+-- @usage
+-- for icell in ipattern:cells() do
+--     print(icell.x, icell.y)
+-- end
 function pattern.cells(ip)
     assert(getmetatable(ip) == pattern, "pattern.cells requires a pattern as the first argument")
-    local icell = 0
-    local ncell = ip:size()
+    local icell, ncells = 0, ip:size()
     return function()
         icell = icell + 1
-        if icell <= ncell then
+        if icell <= ncells then
             local ikey = ip.cellkey[icell]
             local x, y = key_to_coordinates(ikey)
             return cell.new(x,y)
@@ -226,54 +258,81 @@ function pattern.cells(ip)
     end
 end
 
---- Return an iterator over active cell coordinates in the pattern.
--- Simmilar to `pattern.cells` but provides an iterator that runs
--- over (x,y) coordinates instead of `cell` instances. Unlike in
--- `pattern.cells` no tables are created here.
+--- Iterator over active cell coordinates in the pattern.
+-- Simmilar to `pattern.cells` but provides an iterator that runs over (x,y)
+-- coordinates instead of `cell` instances. Normally faster than
+-- `pattern.cells` as no tables are created here.
 -- @param ip source pattern for active cell iterator
--- @return an iterator over cell (x,y) coordinates
+-- @return an iterator returning active cell (x,y) coordinates
+-- @usage
+-- for ix, iy in ipattern:cell_coordinates() do
+--     print(ix, iy)
+-- end
 function pattern.cell_coordinates(ip)
     assert(getmetatable(ip) == pattern, "pattern.cell_coordinates requires a pattern as the first argument")
-    local icell = 0
-    local ncell = ip:size()
+    local icell, ncells = 0, ip:size()
     return function()
         icell = icell + 1
-        if icell <= ncell then
+        if icell <= ncells then
             local ikey = ip.cellkey[icell]
             return key_to_coordinates(ikey)
         end
     end
 end
 
---- Return an iterator over active cells in the pattern in a random order.
--- Simmilar to `pattern.cells` but provides an iterator that returns cells in a randomised order.
+
+--- Shuffled iterator over active cells in the pattern.
+-- Simmilar to `pattern.cells` but provides an iterator that returns cells in a
+-- randomised order, according to a provided random number generator. See
+-- `pattern.cells` for usage.
 -- @param ip source pattern for active cell iterator
 -- @param rng (optional) A random number generating table, following the signature of math.random
--- @return an iterator providing a `cell` table for every active cell in the pattern, in a randomised order
+-- @return an iterator returning a `cell` for every active cell in the pattern, in a randomised order
 function pattern.shuffled_cells(ip, rng)
-    assert(getmetatable(ip) == pattern, "pattern.shuffled_cells requires a pattern as the first argument")
+    assert(getmetatable(ip) == pattern,
+           "pattern.shuffled_cells requires a pattern as the first argument")
     if rng == nil then rng = math.random end
-    local icell = 0
-    local ncell = ip:size()
+    local icell, ncells = 0, ip:size()
 
     -- Copy and Fisher-Yates shuffle
     local cellkeys = ip.cellkey
-    local skeys = {}
-    for i = 1, ncell, 1 do
-        local j = rng(1, i)
-        if j ~= i then
-            skeys[i] = skeys[j]
-        end
-        skeys[j] = cellkeys[i]
-    end
+    local skeys = shuffled_copy(cellkeys, rng)
 
     -- Return iterator
     return function()
         icell = icell + 1
-        if icell <= ncell then
+        if icell <= ncells then
             local ikey = skeys[icell]
             local x, y = key_to_coordinates(ikey)
             return cell.new(x,y)
+        end
+    end
+end
+
+--- Shuffled iterator over active cell coordinates in the pattern.
+-- Simmilar to `pattern.cell_coordinates` but returns cell (x,y) coordinates in
+-- a randomised order according to a provided random number generator. See
+-- `pattern.cell_coordinates` for usage.
+-- @param ip source pattern for active cell
+-- iterator
+-- @param rng (optional) A random number generating table, following the signature of math.random
+-- @return an iterator returning active cell (x,y) coordinates, randomly shuffled
+function pattern.shuffled_coordinates(ip, rng)
+    assert(getmetatable(ip) == pattern,
+           "pattern.shuffled_coordinates requires a pattern as the first argument")
+    if rng == nil then rng = math.random end
+    local icell, ncells = 0, ip:size()
+
+    -- Copy and Fisher-Yates shuffle
+    local cellkeys = ip.cellkey
+    local skeys = shuffled_copy(cellkeys, rng)
+
+    -- Return iterator
+    return function()
+        icell = icell + 1
+        if icell <= ncells then
+            local ikey = skeys[icell]
+            return key_to_coordinates(ikey)
         end
     end
 end
@@ -288,15 +347,14 @@ end
 -- @param ip The pattern to be rendered as a string
 -- @return pattern as string
 function pattern.__tostring(ip)
-    local string = ''
+    local string = '- pattern origin: ' .. tostring(ip.min) .. '\n'
     for y = ip.min.y, ip.max.y, 1 do
-        string = string
         for x = ip.min.x, ip.max.x, 1 do
-            string = string .. (ip:has_cell(x,y) and ip.onchar or ip.offchar)
+            local char = ip:has_cell(x,y) and ip.onchar or ip.offchar
+            string = string .. char
         end
         string = string .. '\n'
     end
-
     return string
 end
 
@@ -323,8 +381,8 @@ end
 -- @param b pattern to be subtracted from a
 -- @return New pattern consisting of the subset of cells in a which are not in b
 function pattern.__sub(a,b)
-    assert(getmetatable(a) == pattern, "pattern addition requires a pattern as the first argument")
-    assert(getmetatable(b) == pattern, "pattern addition requires a pattern as the second argument")
+    assert(getmetatable(a) == pattern, "pattern subtraction requires a pattern as the first argument")
+    assert(getmetatable(b) == pattern, "pattern subtraction requires a pattern as the second argument")
 
     local c = pattern.new()
     for x,y in a:cell_coordinates() do
@@ -495,15 +553,27 @@ function pattern.enlarge(ip, f)
     return ep
 end
 
+--- Rotate a pattern by 90° clockwise about the origin
+-- @param ip pattern to be rotated
+-- @return copy of `ip` which has been rotated by 90°
+function pattern.rotate(ip)
+    assert(getmetatable(ip) == pattern, "pattern.rotate requires a pattern as the first argument")
+    local np = pattern.new()
+    for x, y in ip:cell_coordinates() do
+        np:insert(y, -x)
+    end
+    return np
+end
+
 --- Generate a copy of a pattern, mirroring it vertically.
 -- @param ip pattern for reflection
 -- @return copy of `ip` which has been is reflected vertically
 function pattern.vreflect(ip)
     assert(getmetatable(ip) == pattern, "pattern.vreflect requires a pattern as the first argument")
     local np = pattern.clone(ip)
-    for v in ip:cells() do
-        local new_y = 2*ip.max.y - v.y + 1
-        np:insert(v.x, new_y)
+    for vx, vy in ip:cell_coordinates() do
+        local new_y = 2*ip.max.y - vy + 1
+        np:insert(vx, new_y)
     end
     return np
 end
@@ -515,29 +585,35 @@ end
 function pattern.hreflect(ip)
     assert(getmetatable(ip) == pattern, "pattern.hreflect requires a pattern as the first argument")
     local np = pattern.clone(ip)
-    for v in ip:cells() do
-        local new_x = 2*ip.max.x - v.x + 1
-        np:insert(new_x, v.y)
+    for vx, vy in ip:cell_coordinates() do
+        local new_x = 2*ip.max.x - vx + 1
+        np:insert(new_x, vy)
     end
     return np
 end
 
---- Generate a pattern consisting of edge cells to a provided pattern.
--- Note that this will *not* necessarily generate a hull, it just returns the
--- inactive neighbours of the provided pattern.
+--- Generate a pattern consisting of all cells on the edge of a provided pattern.
+-- This returns a new pattern consisting of the inactive neighbours of an input
+-- pattern, for a given definition of neighbourhood. Therefore the `edge`
+-- method is useful for either enlarging patterns along their surface, or
+-- determining a *border* of a pattern that does not overlap with the pattern
+-- itself.
 -- @param ip pattern for which the edges should be calculated
 -- @param nbh defines which neighbourhood to scan in to determine edges (default 8/moore)
--- @return A pattern representing the edge of ip
+-- @return A pattern representing the edge (outer hull) of ip
 function pattern.edge(ip, nbh)
-    assert(getmetatable(ip) == pattern, "pattern.edge requires a pattern as the first argument")
     local ep = pattern.new()
     nbh = nbh or neighbourhood.moore()
-    for v in ip:cells() do
+    assert(getmetatable(ip) == pattern, "pattern.edge requires a pattern as the first argument")
+    assert(getmetatable(nbh) == neighbourhood, "pattern.edge requires a neighbourhood as an argument")
+    for ix, iy in ip:cell_coordinates() do
         for j=1, #nbh, 1 do
-            local vpr = v + nbh[j]
-            if ip:has_cell(vpr.x, vpr.y) == false then
-                if ep:has_cell(vpr.x, vpr.y) == false then
-                    ep:insert(vpr.x, vpr.y)
+            local jnbh = nbh[j]
+            local vx = ix + jnbh.x
+            local vy = iy + jnbh.y
+            if ip:has_cell(vx, vy) == false then
+                if ep:has_cell(vx, vy) == false then
+                    ep:insert(vx, vy)
                 end
             end
         end
@@ -546,20 +622,26 @@ function pattern.edge(ip, nbh)
 end
 
 --- Generate a pattern consisting of cells on the surface of a provided pattern.
--- This is simmilar to pattern.edge, but will return cells that are /internal/
--- to the provided pattern.
+-- This returns a new pattern consisting of all active cells in an input pattern
+-- that *neighbour* inactive cells. It is therefore very simmilar to `pattern.edge` but
+-- returns a pattern which completely overlaps with the input pattern. This is therefore
+-- useful when *shrinking* a pattern by removing a cell from its surface, or determining
+-- a *border* of a pattern which consists of cells that are present in the original pattern.
 -- @param ip pattern for which the surface should be calculated
--- @param nbh defines which neighbourhood to scan in to determine edges (default 8/moore)
--- @return A pattern representing the surface of ip
+-- @param nbh defines which neighbourhood to scan in to determine the surface (default 8/moore)
+-- @return A pattern representing the surface (inner hull) of ip
 function pattern.surface(ip, nbh)
-    assert(getmetatable(ip) == pattern, "pattern.edge requires a pattern as the first argument")
     local sp = pattern.new()
     nbh = nbh or neighbourhood.moore()
-    for v in ip:cells() do
+    assert(getmetatable(ip) == pattern, "pattern.surface requires a pattern as the first argument")
+    assert(getmetatable(nbh) == neighbourhood, "pattern.surface requires a neighbourhood as an argument")
+    for ix, iy in ip:cell_coordinates() do
         for j=1, #nbh, 1 do
-            local vpr = v + nbh[j]
-            if ip:has_cell(vpr.x, vpr.y) == false then
-                sp:insert(v.x, v.y)
+            local jnbh = nbh[j]
+            local vx = ix + jnbh.x
+            local vy = iy + jnbh.y
+            if ip:has_cell(vx, vy) == false then
+                sp:insert(ix, iy)
                 break
             end
         end