+++ tests/neighbourhood.lua
@@ -8,27 +8,40 @@ testNeighbourhood = {}
 function testNeighbourhood:setUp()
     self.pattern_1 = primitives.square(1)
     self.pattern_2 = primitives.square(3)
+    self.pattern_3 = primitives.square(5)
 end
 
--- There should be 2^n categories for a neighbourhood with n elements
-function testNeighbourhood:testVonNeumann()
-     local von_neumann    = neighbourhood.von_neumann()
-     lu.assertEquals(4, #von_neumann )
-     lu.assertEquals(16, #von_neumann.categories )
+-- Tests a generic neighbourhood with `nelm` elements, for which `pmax` is a
+-- pattern with a medoid cell with a completely filled neighbourhood.
+function testNeighbourhood:commonTest(nbh, nelm, pmax)
+     -- There should be 2^n categories for a neighbourhood with n elements.
+     local ncat = math.pow(2, nelm)
+     lu.assertEquals(nelm, #nbh )
+     lu.assertEquals(ncat, nbh:get_ncategories() )
      -- Test categorisation
-     local ct1 = von_neumann:categorise(self.pattern_1, self.pattern_1:medoid())
-     local ct2 = von_neumann:categorise(self.pattern_2, self.pattern_2:medoid())
-     lu.assertEquals(16,ct1) -- Lowest category (single cell)
-     lu.assertEquals(1, ct2) -- Highest category (full neighbourhood)
+     local ctmin = nbh:categorise(self.pattern_1, self.pattern_1:medoid())
+     local ctmax = nbh:categorise(pmax, pmax:medoid())
+     lu.assertEquals(ncat, ctmin) -- Lowest category (single cell)
+     lu.assertEquals(   1, ctmax) -- Highest category (full neighbourhood)
+end
+
+function testNeighbourhood:testVonNeumann()
+     local von_neumann = neighbourhood.von_neumann()
+     self:commonTest(von_neumann, 4, self.pattern_2)
 end
 
 function testNeighbourhood:testMoore()
-     local moore            = neighbourhood.moore()
-     lu.assertEquals(8, #moore)
-     lu.assertEquals(256, #moore.categories)
-     -- Test categorisation
-     local ct1 = moore:categorise(self.pattern_1, self.pattern_1:medoid())
-     local ct2 = moore:categorise(self.pattern_2, self.pattern_2:medoid())
-     lu.assertEquals(256,ct1) -- Lowest category (single cell)
-     lu.assertEquals(1, ct2)  -- Highest category (full neighbourhood)
+     local moore = neighbourhood.moore()
+     self:commonTest(moore, 8, self.pattern_2)
 end
+
+function testNeighbourhood:testDiagonal()
+     local diagonal = neighbourhood.diagonal()
+     self:commonTest(diagonal, 4, self.pattern_2)
+end
+
+function testNeighbourhood:testDiagonal2()
+     local diagonal_2 = neighbourhood.diagonal_2()
+     self:commonTest(diagonal_2, 4, self.pattern_3)
+end
+