+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 ---
 - Initial release
 
-0.2b
----------
+0.2
+---
 
-### Features
+# Features
 - Implemented a 'mask' subpattern that masks out cells according to a provided
   function
 - Added Voronoi relaxation via Lloyd's algorithm
@@ -16,23 +16,47 @@
 - Added a Poisson-disc sampling subpattern
 - Added Mitchell's Best-Candidate sampling (approximate Poisson-Disc)
 
-### Bugfix
+# Bugfix
 - Fixed bug with subpattern.enclosed which would generate more than one enclosed
   point for a primitive.circle
-- Fixed default initialisation of RNG in automata.async_iterate
+- Fixed default initialisation of RNG in `automata.async_iterate`
 
-### Misc
+# Misc
 - Improved circle raster (no longer repeats points)
 - Integrated all tests into a single test script, and added luacov coverage
 - Converted subpattern.random to take as an argument a fixed integer number of
   desired samples rather than a fraction of the domain size.
 - Removed special handling of '-0' coordinate in cell: No longer required with
   integer spatial hash in patterns.
-- Made pattern coordinate limits explicit in MAX_COORDINATE
+- Made pattern coordinate limits explicit in `MAX_COORDINATE`
 - Changed internal structure of `pattern`, from a list of cells to a list of
   coordinate hashes
 - Various optimisations
 - Removed some (confusing) functionality from `cell`, namely addition and
   multiplication with a number value.
 - Added isoline drawing example
-- Renamed pretty_print to print_patterns
+- Renamed `pretty_print` to `print_patterns`
+
+0.3
+---
+
+# Features
+- Pattern rotation operator
+- Shuffled version of `cell_coordinates` iterator
+- Generalised pattern prototype constructor to allow `NxM` matrices 
+
+# Bugfix
+- Fixed lazy initialisation through 'require ("forma")'
+- Corrected assert error message in pattern subtraction
+
+# Misc
+- Improved pattern documentation
+- Greatly expanded test coverage
+- Slightly streamlined corridors example
+- Much faster `subpattern.floodfill`
+- Much faster construction of neighbourhoods
+- Slightly faster pattern surface/enlarge/reflect
+- Slightly faster `subpattern.random`
+- Slightly faster convergence check for `automata.iterate`
+- Harmonised coordinate handling between pattern prototype and tostring
+